requirements.txt
aws-cdk-lib==2.131.0
constructs>=10.0.0,<11.0.0
boto3==1.34.53
botocore==1.34.53
websockets==12.0
opensearch-py==2.4.2
requests==2.31.0
requests-aws4auth==1.2.3